requirements.txt
crispy-bootstrap5==2025.6
Django==5.2.3
django-crispy-forms==2.4
django-filter==25.1
gunicorn==23.0.0
psycopg2-binary==2.9.10
ruff
isort
coverage
